update(modularization)

+++ 2cpp/ex02/Fixed.cpp
@@ -140,20 +140,20 @@ Fixed	Fixed::operator++(int)
 
 Fixed&		Fixed::min(Fixed &n1, Fixed &n2)
 {
-	return ((n1.toFloat() < n2.toFloat()) ? n1 : n2);
+	return ((n1 < n2) ? n1 : n2);
 }
 
 const Fixed	Fixed::min(const Fixed &n1, const Fixed &n2)
 {
-	return ((n1.toFloat() < n2.toFloat()) ? n1 : n2);
+	return ((n1 < n2) ? n1 : n2);
 }
 
 Fixed&		Fixed::max(Fixed &n1, Fixed &n2)
 {
-	return ((n1.toFloat() > n2.toFloat()) ? n1 : n2);
+	return ((n1 > n2) ? n1 : n2);
 }
 
 const Fixed	Fixed::max(const Fixed &n1, const Fixed &n2)
 {
-	return ((n1.toFloat() > n2.toFloat()) ? n1 : n2);
+	return ((n1 > n2) ? n1 : n2);
 }
